refactor(test): move helper func in testsupport pkg to avoid duplication
fixes #405
