+++ pkg/parser/file_inclusion_test.go
@@ -1,19 +1,15 @@
 package parser_test
 
 import (
-	"bytes"
-	"encoding/json"
-	"os"
 	"strings"
 
 	"github.com/bytesparadise/libasciidoc/pkg/parser"
 	"github.com/bytesparadise/libasciidoc/pkg/types"
+	"github.com/bytesparadise/libasciidoc/testsupport"
 
 	"github.com/davecgh/go-spew/spew"
 	. "github.com/onsi/ginkgo"
 	. "github.com/onsi/ginkgo/extensions/table"
-	. "github.com/onsi/gomega"
-	log "github.com/sirupsen/logrus"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 )
@@ -870,7 +866,7 @@ include::../../test/includes/chapter-a.adoc[]
 
 		It("should replace with string element if directory does not exist in standalone block", func() {
 			// setup logger to write in a buffer so we can check the output
-			console, reset := configureLogger()
+			console, reset := testsupport.ConfigureLogger()
 			defer reset()
 			source := `include::{unknown}/unknown.adoc[leveloffset=+1]`
 			expected := types.PreflightDocument{
@@ -889,13 +885,13 @@ include::../../test/includes/chapter-a.adoc[]
 			}
 			verifyPreflight("foo.adoc", expected, source)
 			// verify error in logs
-			verifyConsoleOutput(console, "failed to include '{unknown}/unknown.adoc'")
+			testsupport.VerifyConsoleOutput(console, "failed to include '{unknown}/unknown.adoc'")
 
 		})
 
 		It("should replace with string element if file is missing in standalone block", func() {
 			// setup logger to write in a buffer so we can check the output
-			console, reset := configureLogger()
+			console, reset := testsupport.ConfigureLogger()
 			defer reset()
 
 			source := `include::../../test/includes/unknown.adoc[leveloffset=+1]`
@@ -915,12 +911,12 @@ include::../../test/includes/chapter-a.adoc[]
 			}
 			verifyPreflight("foo.adoc", expected, source)
 			// verify error in logs
-			verifyConsoleOutput(console, "failed to include '../../test/includes/unknown.adoc'")
+			testsupport.VerifyConsoleOutput(console, "failed to include '../../test/includes/unknown.adoc'")
 		})
 
 		It("should replace with string element if file is missing in delimited block", func() {
 			// setup logger to write in a buffer so we can check the output
-			console, reset := configureLogger()
+			console, reset := testsupport.ConfigureLogger()
 			defer reset()
 
 			source := `----
@@ -948,7 +944,7 @@ include::../../test/includes/unknown.adoc[leveloffset=+1]
 			}
 			verifyPreflight("foo.adoc", expected, source)
 			// verify error in logs
-			verifyConsoleOutput(console, "failed to include '../../test/includes/unknown.adoc'")
+			testsupport.VerifyConsoleOutput(console, "failed to include '../../test/includes/unknown.adoc'")
 		})
 	})
 
@@ -1612,31 +1608,3 @@ include::../../test/includes/chapter-a.adoc[]
 		})
 	})
 })
-
-func verifyConsoleOutput(console Readable, errorMsg string) {
-	GinkgoT().Logf(console.String())
-	out := make(map[string]interface{})
-	err := json.Unmarshal(console.Bytes(), &out)
-	Expect(err).ShouldNot(HaveOccurred())
-	Expect(out["level"]).Should(Equal("error"))
-	Expect(out["msg"]).Should(Equal(errorMsg))
-}
-
-func configureLogger() (Readable, func()) {
-	fmtr := log.StandardLogger().Formatter
-
-	buf := bytes.NewBuffer(nil)
-	log.SetOutput(buf)
-	log.SetFormatter(&log.JSONFormatter{
-		DisableTimestamp: true,
-	})
-	return buf, func() {
-		log.SetOutput(os.Stdout)
-		log.SetFormatter(fmtr)
-	}
-}
-
-type Readable interface {
-	Bytes() []byte
-	String() string
-}
