+++ pkg/renderer/html5/html5_test.go
@@ -3,20 +3,16 @@ package html5_test
 import (
 	"bytes"
 	"context"
-	"encoding/json"
 	"strings"
-	"os"
 
 	"github.com/bytesparadise/libasciidoc/pkg/parser"
 	"github.com/bytesparadise/libasciidoc/pkg/renderer"
 	"github.com/bytesparadise/libasciidoc/pkg/renderer/html5"
 	"github.com/davecgh/go-spew/spew"
 	. "github.com/onsi/ginkgo"
-	. "github.com/onsi/gomega"
 	"github.com/sergi/go-diff/diffmatchpatch"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
-	log "github.com/sirupsen/logrus"
 )
 
 func verify(filename, expected, content string, rendererOpts ...renderer.Option) {
@@ -41,32 +37,4 @@ func verify(filename, expected, content string, rendererOpts ...renderer.Option)
 	dmp := diffmatchpatch.New()
 	diffs := dmp.DiffMain(result, expected, true)
 	assert.Equal(GinkgoT(), expected, result, dmp.DiffPrettyText(diffs))
-}
-
-func verifyConsoleOutput(console Readable, errorMsg string) {
-	GinkgoT().Logf(console.String())
-	out := make(map[string]interface{})
-	err := json.Unmarshal(console.Bytes(), &out)
-	Expect(err).ShouldNot(HaveOccurred())
-	Expect(out["level"]).Should(Equal("error"))
-	Expect(out["msg"]).Should(Equal(errorMsg))
-}
-
-func configureLogger() (Readable, func()) {
-	fmtr := log.StandardLogger().Formatter
-
-	buf := bytes.NewBuffer(nil)
-	log.SetOutput(buf)
-	log.SetFormatter(&log.JSONFormatter{
-		DisableTimestamp: true,
-	})
-	return buf, func() {
-		log.SetOutput(os.Stdout)
-		log.SetFormatter(fmtr)
-	}
-}
-
-type Readable interface {
-	Bytes() []byte
-	String() string
 }
